+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/parsing/parsed_document_builder.rb
@@ -22,6 +22,8 @@ class RubyLsp::Ree::ParsedDocumentBuilder
     case type
     when :enum
       build_enum_document(ast)
+    when :dao
+      build_dao_document(ast)
     else
       build_regular_document(ast)
     end
@@ -44,6 +46,15 @@ class RubyLsp::Ree::ParsedDocumentBuilder
     
     document.parse_class_node
     document.parse_values
+
+    document
+  end   
+
+  def self.build_dao_document(ast)
+    document = RubyLsp::Ree::ParsedDocument.new(ast)
+    
+    document.parse_class_node
+    document.parse_filters
 
     document
   end